requirements.txt
pdfplumber
openai
pydantic
polars
seaborn
matplotlib
ipykernel
pyarrow
scikit-learn
# pytesseract
google-cloud-vision